refactor(vulinbox agent client): don't depend on package vulinbox
