+++ .github/workflows/essential-tests.yml
@@ -94,4 +94,4 @@ jobs:
         run: go test -timeout 1m -run TestMUSTPASS ./common/utils -v
 
       - name: Test Vulinbox agent client
-        run: go test -timeout 1m ./common/vulinboxAgentClient -v
+        run: go test -timeout 1m ./common/vulinboxagentclient -v
