+++ common/vulinboxagentclient/options.go
@@ -1,4 +1,4 @@
-package vulinboxAgentClient
+package vulinboxagentclient
 
 func WithOnClose(f func()) Option {
 	return func(c *Client) {
