+++ common/vulinboxagentproto/proto.go
@@ -1,10 +1,8 @@
-package vulinbox
+package vulinboxagentproto
 
 import (
 	"encoding/base64"
-	"github.com/yaklang/yaklang/common/utils"
 	"math/rand"
-	"net"
 	"net/netip"
 	"time"
 )
@@ -97,44 +95,4 @@ type UnsubscribeAction struct {
 	AgentProtocol
 	Type  string   `json:"type"`
 	Rules []string `json:"rules"`
-}
-
-func handlePing(_ []byte) (any, error) {
-	return nil, nil
-}
-
-func handleUDP(data []byte) (any, error) {
-	udp := utils.MustUnmarshalJson[UDPAction](data)
-	if udp == nil {
-		return nil, nil
-	}
-	conn, err := net.DialUDP("udp", nil, net.UDPAddrFromAddrPort(udp.Target))
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-
-	bytes, err := base64.StdEncoding.DecodeString(udp.Content)
-	if err != nil {
-		return nil, err
-	}
-
-	if _, err = conn.Write(bytes); err != nil {
-		return nil, err
-	}
-
-	if udp.WaitTimeout == 0 {
-		return nil, nil
-	}
-
-	if err := conn.SetDeadline(time.Now().Add(udp.WaitTimeout)); err != nil {
-		return nil, err
-	}
-
-	buf := make([]byte, 1024)
-	_, _, err = conn.ReadFromUDP(buf)
-	if err != nil {
-		return nil, err
-	}
-	return buf, nil
 }
